servicio-2/pyproject.toml
[project]
name = "servicio-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "influxdb-client>=1.48.0",
    "paho-mqtt>=2.1.0",
]
